Add min and Max values to the voltage graph
They can be specified with -m (min) and -M (Max)
The arguments don't change any recordings, just the display of the
voltage graph if GUI is enabled. Any values above or below limits
are drawn as red, while safe values are still blue

This is done by drawing two lines on top of eachother with
identical X values. The top is the red line and it's Y values are
set to `NaN` when the voltage is within the threshold, Plots.jl
inperprets this value as invisible so the blue line below is
shown. This does require generateing aditional points along both
lines when the threshold is crossed, so the graph line doesn't
continue across to the next point. Each edge case of the line
crossing a threshold requires creating specific points.

Because we are showing many graphs at once with many different
points this causes a panic in Plots.jl which is fixed by adding
duplicate points to the other graphs.